Only require html report templates when generating html report
Reading html template files from disk when not generating html-reports is wasteful. It's also a lot easier to configure a bundler for a serverless deployment if the runtime isn't reading files on disk that are not in the module dendendency graph.

Case in point: I'm bundling up my lighthouse with webpack to minimize my lambda deployment package, but am running into errors because these template files are not included in the dependency graph. I'm only generating a json report, so I really don't need them